python_service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
firebase-admin>=6.2.0
pydantic>=2.4.2
httpx>=0.25.0
python-multipart>=0.0.6
pytest>=7.4.3
moviepy>=1.0.3
pillow>=10.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydub>=0.25.1
